TEIIDDES-1530 Added method to clean up incompatible MEDS from a copied model.

+++ plugins/org.teiid.designer.core/src/org/teiid/designer/core/extension/ModelExtensionUtils.java
@@ -600,7 +600,7 @@ class ModelExtensionUtils {
         if (annotation != null) {
             annotation.getTags().remove(propId);
 
-            if (deleteModelObjectAnnotation && annotation.getTags().isEmpty()) {
+            if (deleteModelObjectAnnotation && annotation.getTags().isEmpty() && annotation.getDescription() == null) {
                 ModelResourceContainerFactory.deleteAnnotation(annotation);
             }
         }
